+++ bootique-job-consul/src/main/java/io/bootique/job/consul/ConsulJobModule.java
@@ -29,7 +29,7 @@ import io.bootique.job.lock.LockHandler;
 import io.bootique.shutdown.ShutdownManager;
 
 /**
- * @since 0.26
+ * @since 1.0.RC1
  */
 public class ConsulJobModule extends ConfigModule {