+++ bootique-job-consul/src/main/java/io/bootique/job/consul/ConsulJobModuleProvider.java
@@ -26,7 +26,7 @@ import java.util.Collection;
 import java.util.Collections;
 
 /**
- * @since 0.26
+ * @since 1.0.RC1
  */
 public class ConsulJobModuleProvider implements BQModuleProvider {