+++ bootique-job-consul/src/main/java/io/bootique/job/consul/lock/ConsulLockHandler.java
@@ -30,7 +30,7 @@ import org.slf4j.LoggerFactory;
 import java.util.function.Supplier;
 
 /**
- * @since 0.26
+ * @since 1.0.RC1
  */
 public class ConsulLockHandler implements LockHandler {